Add ErrorHandlerManager
